% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_gene.R
\name{nb_gene}
\alias{nb_gene}
\title{Estimate Burst Parameters of a Single Gene from Count Data}
\usage{
nb_gene(vals, depths = NULL)
}
\arguments{
\item{vals}{The vector of raw counts for a gene.}

\item{depths}{Optional numeric vector giving the relative sequencing depth
of each cell. If \code{NULL}, sequencing depth is assumed equal across 
cells.}
}
\description{
This function takes a vector of gene counts and estimates burst parameters 
by fitting a negative binomial (NB) model. The NB size parameter \code{r} 
and success probability \code{p} are estimated, and sequencing depths
corresponding to each value can optionally be included.
}
